chore: Restructure packages (#61)
* restructure

* lint

* test restructure+workflows

* lint

* add temporary workaround

* remove log

* change tsconfig setup

* mock env in global setup

* fix tsconfig.test

* fix test

* add core to canary release

* add ignored packages to config.json

* Update test-util/mock-http.ts


* Update packages/core/src/context.test.ts


* review

* fix: Changes from lint

* update docs

* docs

* fix: Changes from lint

---------
