+++ .github/workflows/build.yml
@@ -25,7 +25,7 @@ jobs:
           node-version: ${{ matrix.node-version }}
           cache: 'pnpm'
       - run: pnpm i --frozen-lockfile
-      - run: pnpm test
+      - run: pnpm test:unit
       - run: pnpm test:type
 
   checks:
