+++ .github/workflows/e2e-test.yaml
@@ -52,9 +52,7 @@ jobs:
           done
           wget -qO- -S --content-on-error localhost:8080
       - name: "Execute E2E Tests"
-        working-directory: ./tests/e2e-tests
-        run: |
-          pnpm run e2e-test
+        run: pnpm test:e2e
       - name: "Slack Notification"
         if: failure()
         uses: slackapi/slack-github-action@v1.26.0
