+++ eslint.config.js
@@ -3,6 +3,11 @@ import flatConfig from '@sap-cloud-sdk/eslint-config/flat-config.js';
 export default [
   ...flatConfig,
   {
+    // Estlint flat config is not supported by eslint-plugin-import.
+    // https://github.com/import-js/eslint-plugin-import/issues/2556
+    rules: { 'import/namespace': 'off'}
+  },
+  {
     ignores: ['**/dist/**/*', '**/coverage/**/*', 'packages/ai-core/src/client/**/*'],
   },
   {
