+++ jest.config.mjs
@@ -54,10 +54,10 @@ const config = {
   // forceCoverageMatch: [],
 
   // A path to a module which exports an async function that is triggered once before all test suites
-  // globalSetup: undefined,
+   globalSetup: '../../global-test-setup.ts',
 
   // A path to a module which exports an async function that is triggered once after all test suites
-  // globalTeardown: undefined,
+   globalTeardown: '../../global-test-teardown.ts',
 
   // A set of global variables that need to be available in all test environments
   // globals: {},
@@ -174,6 +174,7 @@ const config = {
       'ts-jest',
       {
         useESM: true,
+        tsconfig: '../../tsconfig.test.json',
       },
     ],
   },
