+++ package.json
@@ -7,24 +7,25 @@
   "repository": "github:SAP/ai-sdk-js",
   "private": true,
   "type": "module",
-  "types": "tests/type-tests",
-  "tsd": {
-    "directory": "tests/type-tests"
-  },
   "scripts": {
     "postinstall": "pnpm compile",
     "compile": "pnpm -r -w=false run compile",
-    "test": "NODE_OPTIONS=--experimental-vm-modules pnpm -r run test",
-    "test:type": "tsd",
+    "test:unit": "pnpm -r -F=./packages/** test",
+    "test:type": "pnpm -F=@sap-ai-sdk/type-tests test",
+    "test:e2e": "pnpm -F=@sap-ai-sdk/e2e-tests test",
     "lint": "pnpm -r run lint",
     "lint:fix": "pnpm -r run lint:fix",
-    "generate": "pnpm -r run generate"
+    "generate": "pnpm -r run generate",
+    "ai-core": "pnpm -F=@sap-ai-sdk/ai-core",
+    "gen-ai-hub": "pnpm -F=@sap-ai-sdk/gen-ai-hub",
+    "core": "pnpm -F=@sap-ai-sdk/core"
   },
   "devDependencies": {
     "@changesets/cli": "^2.27.7",
     "@sap-cloud-sdk/eslint-config": "^3.18.0",
     "@sap-cloud-sdk/connectivity": "^3.18.0",
     "@sap-cloud-sdk/http-client": "^3.18.0",
+    "@sap-ai-sdk/core": "workspace:^",
     "@types/jest": "^29.5.12",
     "@types/node": "^20.14.15",
     "@jest/globals": "^29.5.12",
@@ -34,7 +35,8 @@
     "prettier": "^3.3.3",
     "ts-jest": "^29.2.4",
     "ts-node": "^10.9.2",
-    "tsd": "^0.31.0",
-    "typescript": "^5.5.4"
+    "typescript": "^5.5.4",
+    "jsonwebtoken": "^9.0.2",
+    "@types/jsonwebtoken": "^9.0.2"
   }
 }
