+++ packages/core/README.md
@@ -1,0 +1,18 @@
+# @sap-ai-sdk/core
+
+This package contains core utility functions that we reuse in the SDK to set the context and execute HTTP requests.
+They are primarily designed for internal usage.
+
+### Installation
+
+```
+$ npm install @sap-ai-sdk/core
+```
+
+### Usage
+
+The core package is not intended for direct usage.
+
+## License
+
+The SAP AI SDK is released under the [Apache License Version 2.0.](http://www.apache.org/licenses/)
