+++ packages/core/jest.config.mjs
@@ -1,5 +1,5 @@
 import config from '../../jest.config.mjs';
 export default {
   ...config,
-  displayName: 'ai-core',
+  displayName: 'core',
 };
