+++ packages/core/package.json
@@ -1,13 +1,11 @@
 {
-  "name": "@sap-ai-sdk/gen-ai-hub",
+  "name": "@sap-ai-sdk/core",
   "version": "0.0.0",
   "description": "",
   "license": "Apache-2.0",
   "keywords": [
     "sap-ai-sdk",
-    "gen-ai-hub",
-    "orchestration",
-    "llm-access"
+    "ai-core"
   ],
   "type": "module",
   "main": "./dist/index.js",
@@ -22,20 +20,14 @@
     "compile": "tsc",
     "test": "NODE_OPTIONS=--experimental-vm-modules jest",
     "lint": "eslint . && prettier . --config ../../.prettierrc --ignore-path ../../.prettierignore -c",
-    "lint:fix": "eslint . --fix && prettier . --config ../../.prettierrc --ignore-path ../../.prettierignore -w --log-level error",
-    "generate": "pnpm generate:orchestration",
-    "generate:orchestration": "openapi-generator --overwrite --generateESM -i ./src/orchestration/spec/api.yaml -o ./src/orchestration/client && pnpm lint:fix"
+    "lint:fix": "eslint . --fix && prettier . --config ../../.prettierrc --ignore-path ../../.prettierignore -w --log-level error"
   },
   "dependencies": {
-    "@sap-ai-sdk/core": "workspace:^",
     "@sap-cloud-sdk/http-client": "^3.18.0",
     "@sap-cloud-sdk/connectivity": "^3.18.0",
-    "@sap-cloud-sdk/util": "^3.18.0",
-    "@sap-cloud-sdk/openapi": "^3.18.0"
+    "@sap-cloud-sdk/util": "^3.18.0"
   },
   "devDependencies": {
-    "@sap-cloud-sdk/openapi-generator": "^3.18.0",
-    "typescript": "^5.5.4",
-    "nock": "^13.5.4"
+    "typescript": "^5.5.4"
   }
 }
