+++ packages/core/src/http-client.test.ts
@@ -1,26 +1,18 @@
-import { jest } from '@jest/globals';
-import { HttpDestination } from '@sap-cloud-sdk/connectivity';
-import { mockGetAiCoreDestination } from '../test-util/mock-context.js';
-import { mockInference } from '../test-util/mock-http.js';
-
-jest.unstable_mockModule('./context.js', () => ({
-  getAiCoreDestination: jest.fn(() =>
-    Promise.resolve(mockGetAiCoreDestination())
-  )
-}));
-const { executeRequest } = await import('./http-client.js');
+import nock from 'nock';
+import {
+  mockClientCredentialsGrantCall,
+  mockInference
+} from '../../../test-util/mock-http.js';
+import { dummyToken } from '../../../test-util/mock-jwt.js';
+import { executeRequest } from './http-client.js';
 
 describe('http-client', () => {
-  let destination: HttpDestination;
-
   beforeAll(() => {
-    destination = mockGetAiCoreDestination();
+    mockClientCredentialsGrantCall({ access_token: dummyToken }, 200);
   });
-
-  afterAll(() => {
-    jest.restoreAllMocks();
+  afterEach(() => {
+    nock.cleanAll();
   });
-
   it('should execute a request to the AI Core service', async () => {
     const mockPrompt = { prompt: 'some test prompt' };
     const mockPromptResponse = { completion: 'some test completion' };
@@ -35,8 +27,7 @@ describe('http-client', () => {
       {
         data: mockPromptResponse,
         status: 200
-      },
-      destination
+      }
     );
     const res = await executeRequest(
       { url: '/mock-endpoint', apiVersion: 'mock-api-version' },
