+++ packages/core/src/index.ts
@@ -1,0 +1,7 @@
+export {
+  executeRequest,
+  BaseLlmParametersWithDeploymentId,
+  BaseLlmParameters,
+  CustomRequestConfig
+} from './http-client.js';
+export { getAiCoreDestination } from './context.js';
