+++ packages/gen-ai-hub/src/index.ts
@@ -5,13 +5,12 @@ export {
   OpenAiEmbeddingOutput,
   OpenAiChatCompletionOutput
 } from './client/index.js';
-export { CustomRequestConfig, BaseLlmParameters } from './core/index.js';
 export {
   GenAiHubClient,
   GenAiHubCompletionParameters,
   GenAiHubCompletionResponse,
   PromptConfig,
   LlmConfig,
-  ChatMessages
+  ChatMessages,
+  CompletionPostResponse
 } from './orchestration/index.js';
-export { getAiCoreDestination } from './core/index.js';
