+++ packages/gen-ai-hub/src/orchestration/orchestration-client.test.ts
@@ -1,38 +1,30 @@
 import nock from 'nock';
-import { jest } from '@jest/globals';
-import { HttpDestination } from '@sap-cloud-sdk/connectivity';
-import { mockGetAiCoreDestination } from '../test-util/mock-context.js';
-import { mockInference, parseMockResponse } from '../test-util/mock-http.js';
-import { BaseLlmParametersWithDeploymentId } from '../core/index.js';
+import { BaseLlmParametersWithDeploymentId } from '@sap-ai-sdk/core';
+import {
+  mockClientCredentialsGrantCall,
+  mockInference,
+  parseMockResponse
+} from '../../../../test-util/mock-http.js';
+import { dummyToken } from '../../../../test-util/mock-jwt.js';
 import { CompletionPostResponse } from './client/api/index.js';
 import { GenAiHubCompletionParameters } from './orchestration-types.js';
+import {
+  GenAiHubClient,
+  constructCompletionPostRequest
+} from './orchestration-client.js';
 import { azureContentFilter } from './orchestration-filter-utility.js';
-jest.unstable_mockModule('../core/context.js', () => ({
-  getAiCoreDestination: jest.fn(() =>
-    Promise.resolve(mockGetAiCoreDestination())
-  )
-}));
-
-const { GenAiHubClient, constructCompletionPostRequest } = await import(
-  './orchestration-client.js'
-);
 describe('GenAiHubClient', () => {
-  let destination: HttpDestination;
   const client = new GenAiHubClient();
   const deploymentConfiguration: BaseLlmParametersWithDeploymentId = {
     deploymentId: 'deployment-id'
   };
 
   beforeAll(() => {
-    destination = mockGetAiCoreDestination();
-  });
-
-  afterEach(() => {
-    nock.cleanAll();
+    mockClientCredentialsGrantCall({ access_token: dummyToken }, 200);
   });
 
   afterAll(() => {
-    jest.restoreAllMocks();
+    nock.cleanAll();
   });
 
   it('calls chatCompletion with minimum configuration', async () => {
@@ -63,12 +55,12 @@ describe('GenAiHubClient', () => {
         data: mockResponse,
         status: 200
       },
-      destination,
       {
         url: 'completion'
       }
     );
-    expect(client.chatCompletion(request)).resolves.toEqual(mockResponse);
+    const response = await client.chatCompletion(request);
+    expect(response).toEqual(mockResponse);
   });
 
   it('calls chatCompletion with filter configuration supplied using convenience function', async () => {
@@ -105,7 +97,6 @@ describe('GenAiHubClient', () => {
         data: mockResponse,
         status: 200
       },
-      destination,
       {
         url: 'completion'
       }
@@ -167,7 +158,6 @@ describe('GenAiHubClient', () => {
         data: mockResponse,
         status: 200
       },
-      destination,
       {
         url: 'completion'
       }
@@ -217,12 +207,11 @@ describe('GenAiHubClient', () => {
         data: mockResponse,
         status: 200
       },
-      destination,
       {
         url: 'completion'
       }
     );
-
-    expect(client.chatCompletion(request)).resolves.toEqual(mockResponse);
+    const response = await client.chatCompletion(request);
+    expect(response).toEqual(mockResponse);
   });
 });
