+++ packages/gen-ai-hub/test/openai/openai-chat-completion-success-response.json
