+++ packages/gen-ai-hub/test/openai/openai-embeddings-success-response.json
