+++ packages/gen-ai-hub/test/orchestration/genaihub-chat-completion-success-response.json
