+++ pnpm-workspace.yaml
@@ -2,9 +2,12 @@ packages:
   # all packages in direct subdirs of packages/
   - 'packages/ai-core'
   - 'packages/gen-ai-hub'
+  - 'packages/core'
   # sample code
   - 'sample-code'
   # e2e tests
   - 'tests/e2e-tests'
+  # type tests
+  - 'tests/type-tests'
   # exclude packages that are inside test directories
-  - '!tests/type-tests'
+
