+++ sample-code/src/aiservice.ts
@@ -1,4 +1,3 @@
-// eslint-disable-next-line import/namespace
 import { OpenAiClient } from '@sap-ai-sdk/gen-ai-hub';
 
 const openAiClient = new OpenAiClient();
