+++ tsconfig.test.json
@@ -1,0 +1,9 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "rootDir": ".",
+    "noEmit": true
+  },
+  "include": ["packages/**/*.ts", "test-util/**/*.ts"],
+  "exclude": ["dist/**/*", "node_modules/**/*"]
+}
